Implementaciones_Adicionales

+++ src/main/java/com/Producto/API_Producto/service/ProductosService.java
@@ -1,17 +1,19 @@
 package com.Producto.API_Producto.service;
 
 import com.Producto.API_Producto.model.entities.Productos;
+import com.Producto.API_Producto.model.exceptions.ProductoExistenteException;
 import com.Producto.API_Producto.model.exceptions.ProductoNoEncontradoException;
 import com.Producto.API_Producto.model.interfaces.IProductService;
 import com.Producto.API_Producto.repository.ProductoRepository;
+import com.Producto.API_Producto.util.NormalizerText;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.AutoConfiguration;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
 public class ProductosService implements IProductService {
+
     public final ProductoRepository ProductoRepository;
 
     public ProductosService(ProductoRepository productoRepository) {
@@ -30,12 +32,19 @@ public class ProductosService implements IProductService {
     }
 
     @Override
-    public void saveProducts(Productos productos) {
-        ProductoRepository.save(productos);
+    public Productos saveProducts(Productos productos) {
+        String nombreNormalizado= NormalizerText.normalizar(productos.getNombreProducto());
+        if(ProductoRepository.existsByNombreProducto(nombreNormalizado)){
+            throw new ProductoExistenteException(productos.getNombreProducto());
+        }
+        return ProductoRepository.save(productos);
     }
 
     @Override
     public void deleteProducts(Long id) {
+        if(!ProductoRepository.existsById(id)){
+            throw new ProductoNoEncontradoException(id);
+        }
         ProductoRepository.deleteById(id);
     }
 
@@ -43,10 +52,15 @@ public class ProductosService implements IProductService {
     public Productos updateProducts(Long id,Productos productosNuevos) {
         Productos productoExistente=ProductoRepository.findById(id).orElse(null);
         if(productoExistente==null){
-            return null;
+            throw new ProductoNoEncontradoException(id);
         }
         if(productosNuevos.getNombreProducto()!=null){
-            productoExistente.setNombreProducto(productosNuevos.getNombreProducto());
+            String normalizarNombre=NormalizerText.normalizar(productosNuevos.getNombreProducto());
+            boolean productExists= ProductoRepository.existsByNombreProductoAndIdProducto(normalizarNombre,id);
+            if(productExists){
+                throw new ProductoExistenteException(productosNuevos.getNombreProducto());
+            }
+            productoExistente.setNombreProducto(normalizarNombre);
         }
         if(productosNuevos.getDescripcion()!=null){
             productoExistente.setDescripcion(productosNuevos.getDescripcion());
@@ -55,7 +69,7 @@ public class ProductosService implements IProductService {
             productoExistente.setPrecio(productosNuevos.getPrecio());
         }
         if(productosNuevos.getStock()!=null){
-            productoExistente.setPrecio(productosNuevos.getPrecio());
+            productoExistente.setStock(productosNuevos.getStock());
         }
         if(productosNuevos.getEstadoProducto()!=null){
             productoExistente.setEstadoProducto(productosNuevos.getEstadoProducto());
